After reading some documentation undo an incorrect fix

+++ src/main/java/nom/bdezonia/zorbage/type/float32/real/Float32Algebra.java
@@ -1215,7 +1215,7 @@ public class Float32Algebra
 	{
 		@Override
 		public void call(Float32Member a, Float32Member b) {
-			b.setV(Math.nextAfter(a.v(),Float.NEGATIVE_INFINITY));
+			b.setV( Math.nextAfter( a.v(), Double.NEGATIVE_INFINITY ) ) ; // Note that Double is intentional here
 		}
 	};
 	
@@ -1229,7 +1229,7 @@ public class Float32Algebra
 	{
 		@Override
 		public void call(Float32Member a, Float32Member b) {
-			b.setV(Math.nextUp(a.v()));
+			b.setV( Math.nextUp( a.v() ) );
 		}
 	};
